Don't merge this - CP team optimistic locking testing
This branch should likely not be merged since it hardcodes the the CP team's rancher scratch
test enviroment. We are considering using the karate tests for our test of optimistic locking.

For background on optimistic locking implementation see:
https://issues.folio.org/browse/UXPROD-3089

The reason it should not be merged is that the rancher environment ingress okapi url is
specific to theconsidering using the karate tests for our test of optimistic locking. A better
approach would be to make the environment more configurable so that the endpoints didn't have
to be hardcoded in each module's karate tests.

Created a ticket for this: https://issues.folio.org/projects/FAT/issues/FAT-952